formalize the timestamp_provider mechanism
- initially I thought that the Config object should be passed around
  (which I'm not yet convinced that this shouldnt be the case), but
  making an arbitrary choice here to limit the api surface
- only expose the timestamp_provider methods from emitters
- open to refactoring later, this is not public crate api
