+++ src/output/emitter.rs
@@ -17,8 +17,7 @@ use crate::output::{
 use crate::spec;
 
 pub struct JsonEmitter {
-    // HACK: public for tests, but this should come from config directly to where needed
-    pub(crate) timestamp_provider: Box<dyn config::TimestampProvider + Send + Sync + 'static>,
+    timestamp_provider: Box<dyn config::TimestampProvider + Send + Sync + 'static>,
     writer: writer::WriterType,
     seqno: Arc<atomic::AtomicU64>,
 }
@@ -35,21 +34,26 @@ impl JsonEmitter {
         }
     }
 
-    fn serialize_artifact(&self, object: &spec::RootImpl) -> serde_json::Value {
+    fn incr_seqno(&self) -> u64 {
+        self.seqno.fetch_add(1, Ordering::AcqRel)
+    }
+
+    fn serialize_artifact(&self, object: &spec::RootImpl) -> String {
         let root = spec::Root {
             artifact: object.clone(),
             timestamp: self.timestamp_provider.now(),
             seqno: self.incr_seqno(),
         };
-        serde_json::json!(root)
+
+        serde_json::json!(root).to_string()
     }
 
-    fn incr_seqno(&self) -> u64 {
-        self.seqno.fetch_add(1, Ordering::AcqRel)
+    pub fn timestamp_provider(&self) -> &(dyn config::TimestampProvider + Send + Sync + 'static) {
+        &*self.timestamp_provider
     }
 
     pub async fn emit(&self, object: &spec::RootImpl) -> Result<(), io::Error> {
-        let s = self.serialize_artifact(object).to_string();
+        let s = self.serialize_artifact(object);
 
         match &self.writer {
             WriterType::File(file) => file.write(&s).await?,
@@ -72,6 +76,20 @@ mod tests {
 
     use super::*;
 
+    pub struct NullTimestampProvider {}
+
+    impl NullTimestampProvider {
+        // warn: linter is wrong here, this is used in a serde_json::json! block
+        #[allow(dead_code)]
+        pub const FORMATTED: &str = "1970-01-01T00:00:00.000Z";
+    }
+
+    impl config::TimestampProvider for NullTimestampProvider {
+        fn now(&self) -> chrono::DateTime<chrono_tz::Tz> {
+            chrono::DateTime::from_timestamp_nanos(0).with_timezone(&chrono_tz::UTC)
+        }
+    }
+
     #[tokio::test]
     async fn test_emit_using_buffer_writer() -> Result<()> {
         let expected = json!({
@@ -80,13 +98,13 @@ mod tests {
                 "minor": spec::SPEC_VERSION.1,
             },
             "sequenceNumber": 0,
-            "timestamp": config::NullTimestampProvider::FORMATTED,
+            "timestamp": NullTimestampProvider::FORMATTED,
         });
 
         let buffer = Arc::new(Mutex::new(vec![]));
         let writer = writer::BufferWriter::new(buffer.clone());
         let emitter = JsonEmitter::new(
-            Box::new(config::NullTimestampProvider {}),
+            Box::new(NullTimestampProvider {}),
             writer::WriterType::Buffer(writer),
         );
 
@@ -112,7 +130,7 @@ mod tests {
                 "minor": spec::SPEC_VERSION.1,
             },
             "sequenceNumber": 0,
-            "timestamp": config::NullTimestampProvider::FORMATTED,
+            "timestamp": NullTimestampProvider::FORMATTED,
         });
         let expected_2 = json!({
             "schemaVersion": {
@@ -120,13 +138,13 @@ mod tests {
                 "minor": spec::SPEC_VERSION.1,
             },
             "sequenceNumber": 1,
-            "timestamp": config::NullTimestampProvider::FORMATTED,
+            "timestamp": NullTimestampProvider::FORMATTED,
         });
 
         let buffer = Arc::new(Mutex::new(vec![]));
         let writer = writer::BufferWriter::new(buffer.clone());
         let emitter = JsonEmitter::new(
-            Box::new(config::NullTimestampProvider {}),
+            Box::new(NullTimestampProvider {}),
             writer::WriterType::Buffer(writer),
         );
